// src/models/rnevents/RnMqttEventParams.ts
export enum RnMqttEventParams {
  MQTT_PARAM_RECONNECT = 'MQTT_PARAM_RECONNECT',
  MQTT_PARAM_SERVER_URI = 'MQTT_PARAM_SERVER_URI',
  MQTT_PARAM_CLIENT_REF = 'MQTT_PARAM_CLIENT_REF',
  MQTT_PARAM_ERR_MESSAGE = 'MQTT_PARAM_ERR_MESSAGE',
  MQTT_PARAM_ERR_CODE = 'MQTT_PARAM_ERR_CODE',
  MQTT_PARAM_STACKTRACE = 'MQTT_PARAM_STACKTRACE',
  MQTT_PARAM_TOPIC = 'MQTT_PARAM_TOPIC',
  MQTT_PARAM_MESSAGE = 'MQTT_PARAM_MESSAGE',
  MQTT_PARAM_QOS = 'MQTT_PARAM_QOS',
  MQTT_PARAM_RETAIN = 'MQTT_PARAM_RETAIN',
}